Moved save/restore functions to craft.go.
